Fix lack of uniqueId on AMD GPU OpenCL without AMD extensions
AMD GPU systems need not have a functional
clGetDeviceInfo(..., CL_DEVICE_TOPOLOGY_AMD, ...).

E.g., Fedora OpenCL with AMD GPU drivers, it seems.  As the code stands it
doesn't create a uniqueId, and leaves it null.  If there are multiple GPUs,
the device info is stored to null uniqueId and overwritten - and only one of
these GPUs will be enumerated and visible to the miner.

Fix this by providing a fall-back, constructing the uniqueId from the
platform and device indexes.  As is already the case for Intel GPU and CPU
devices.

Additionally, cosmetically, add a clearer, more self-describing struct for
the AMD topology data.